cs_22-master/cs_22-master/requirements.txt
asgiref==3.9.1
attrs==25.4.0
boto3==1.40.30
botocore==1.40.30
dj-database-url==3.1.0
Django==5.2.6
django-cors-headers==4.9.0
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.27.2
gunicorn==25.0.1
inflection==0.5.1
jmespath==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
packaging==26.0
pillow==11.0.0
PyJWT==2.11.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
qrcode==7.4.2
referencing==0.37.0
rpds-py==0.30.0
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.11.0
